Origin/feat/camera door (#57)
* Camera improved and added physics utils

* enemy animated

* merge fixes

* first clang fixes

* world fix

* more clang fixes

* final clang fixes

* clangity

* boosted ldtk loader with create enemy function

* enemy death, animatios, etc

* feat: enable debug mode

* feat: added circle to player

* exp M attack animation

* Refactored createBullet

* Added changes

* Enemy Bullets

* clang fixes

* fix: fixed compilation issue

* clang fixes

* Delete locator.cpp

* Update components.cmake

* clang

---------

+++ src/missile_engine/src/common.cpp
@@ -21,7 +21,7 @@ std::optional<std::pair<std::unique_ptr<uint8_t[]>, size_t>> missileengine::load
     PHYSFS_sint64 size = PHYSFS_fileLength(file);
 
     // Allocate memory for the file
-    auto data = std::make_unique<uint8_t[]>(size); // NOLINT(*-avoid-c-arrays)
+    auto data = std::make_unique<uint8_t[]>(static_cast<size_t>(size)); // NOLINT(*-avoid-c-arrays)
 
     // Read the file into memory
     if (PHYSFS_readBytes(file, data.get(), size) == -1)
